requirements.txt
pdfplumber==0.10.3
pandas==2.2.1
openpyxl==3.1.2
tqdm==4.66.2
python-dotenv==1.0.1
google-genai==0.3.0
